Rename st8.hpp to stm8.hpp
