+++ ana.cpp
@@ -1,4 +1,4 @@
-#include "st8.hpp"
+#include "stm8.hpp"
 
 #define F_NONDX		0
