+++ reg.cpp
@@ -1,4 +1,4 @@
-#include "st8.hpp"
+#include "stm8.hpp"
 #include <fpro.h>
 #include <diskio.hpp>
 #include <entry.hpp>
