+++ stm8.hpp
